add docs for module subfxns
