+++ README.md
@@ -12,8 +12,8 @@ The components of this app are:
 1. Functions to calculate scores for quality tasks and model predictions.
 2. Custom nodes that can be added to pipelines to calculate scores when a task's quality items are completed.
 
-Currently there is support for quality tasks (i.e qualification, honeypot, and consensus), model prediction scoring, and 
-precision-recall 
+Currently there is support for quality tasks (i.e qualification, honeypot, and consensus), model prediction scoring, and
+precision-recall
 
 ## Python installation
 
@@ -21,10 +21,10 @@ precision-recall
 pip install git+https://github.com/dataloop-ai-apps/dtlpy-metrics
 ```
 
-
 ## Tutorial and How-To
 
-See [this tutorial](../docs/Quality task scoring tutorial.ipynb) for details on scoring for Dataloop "quality tasks" (qualification, honeypot, and consensus tasks).
+See [this tutorial](../docs/Quality task scoring tutorial.ipynb) for details on scoring for Dataloop "quality tasks" (
+qualification, honeypot, and consensus tasks).
 
 ## Functions
 
@@ -36,16 +36,22 @@ See [this tutorial](../docs/Quality task scoring tutorial.ipynb) for details on 
     * Output:
         * `task`: the original task entity
 
-* `calculate_task_item_score`
-    * Description: This function takes items from a consensus task and calculates annotations scores for a given item,
-      including the overall annotation score and overall item score.
+* `create_task_item_score`
+    * Description: This function takes items from a quality task and calculates annotations scores for a given item,
+      including overall annotation scores and the overall item score.
+
+      Scores calculated will vary depending on the item type. For example, images will have individual annotation
+      scores, overall annotation scores, label confusion scores, and user confusion scores. Videos will first calculate
+      scores frame by frame, and then annotation scores will be the average across all frames where the annotation is
+      present.
+
     * Input:
         * `item`: item whose annotations are to be scored
         * `task`: optional task entity for the task that the item belongs to
         * `context`: optional context dictionary for the relevant IDs associated with the item
         * `score_types`: optional list for specifying which scores are to be calculated
     * Output:
-        * `item`
+        * `item`: the original item entity
 
 * `create_model_score`
     * Description: This function takes model predictions (i.e. annotations) and calculates the score for each
@@ -57,24 +63,6 @@ See [this tutorial](../docs/Quality task scoring tutorial.ipynb) for details on 
         * `ignore_labels`: optional boolean for whether to ignore the labels in the model predictions
         * `match_threshold`: optional float for IOU threshold for matching model predictions to ground truth annotations
         * `compare_types`: optional list for specifying which annotation types are to be compared
-
-* `calculate_annotation_scores`
-    * Description: This function takes annotations and calculates the score for each annotation.
-    * Input:
-        * `annot_collection_1`: set 1 of a dl.AnnotationCollection or list of annotations 
-        * `annot_collection_2`: set 2 of a dl.AnnotationCollection or list of annotations
-        * `ignore_labels`: optional boolean for whether to ignore the labels in the model predictions
-        * `match_threshold`: optional float for IOU threshold for matching model predictions to ground truth annotations
-        * `compare_types`: optional list for specifying which annotation types are to be compared
-        * `score_types`: optional list for specifying which scores are to be calculated
-
-* `get_model_scores_df`
-    * Description: This function generates the pandas dataframe of model scores.
-    * Input:
-        * `dataset`: the dataset entity containing the test items that were predicted on
-        * `model`: the model entity that produced the predictions
-    * Output:
-        * `model_scores_df`: pandas dataframe containing the scores
 
 ## Contributions, Bugs and Issues - How to Contribute
