+++ dtlpymetrics/scoring.py
@@ -321,58 +321,3 @@ def create_model_score(dataset: dl.Dataset = None,
                                 overwrite=True)
     logger.info(f'Successfully created model scores and saved as item {item.id}.')
     return model
-
-
-def calculate_confusion_matrix_item(item: dl.Item,
-                                    scores: List[Score],
-                                    save_plot=True) -> pd.DataFrame:
-    """
-    Calculate confusion matrix from a set of label confusion scores
-
-    :return:
-    """
-    scores_dl = []
-    for score in scores:
-        scores_dl.append(Score.from_json(score))
-
-    # ###############################
-    # # create table of comparisons #
-    # ###############################
-    label_names = []
-    for score in scores_dl:
-        if score.type == ScoreType.LABEL_CONFUSION:
-            if score.entity_id not in label_names:
-                label_names.append(score.entity_id)
-            if score.relative not in label_names:
-                label_names.append(score.relative)
-
-    conf_matrix = pd.DataFrame(index=label_names, columns=label_names)
-
-    for score in scores_dl:
-        if score.type == ScoreType.LABEL_CONFUSION:
-            conf_matrix.loc[score.entity_id] = score.value
-            conf_matrix.loc[score.entity_id, score.relative] = score.value
-
-    conf_matrix.fillna(0, inplace=True)
-    conf_matrix.rename(columns={None: 'unlabeled'}, inplace=True)
-    conf_matrix.rename(index={None: 'unlabeled'}, inplace=True)
-    label_names = ['unlabeled' if label is None else label for label in label_names]
-
-    if save_plot is True:
-        if os.environ.get('SCORES_DEBUG_PATH', None) is not None:
-            debug_path = os.environ.get('SCORES_DEBUG_PATH', None)
-
-            plot_matrix(item_title=f'label confusion matrix {item.id}',
-                        filename=os.path.join(debug_path, 'label_confusion',
-                                              f'label_confusion_matrix_{item.id}.png'),
-                        matrix_to_plot=conf_matrix,
-                        axis_labels=label_names)
-
-        else:
-            plot_matrix(item_title=f'label confusion matrix {item.id}',
-                        filename=os.path.join('.dataloop', 'label_confusion',
-                                              f'label_confusion_matrix_{item.id}.png'),
-                        matrix_to_plot=conf_matrix,
-                        axis_labels=label_names)
-
-    return conf_matrix
